BigBuckSensingNode: Added wifi power up timer.
After a certain amount of time with the wifi up, the pi will turn on its wifi
to send to the master its status (this would work well in the case where
the master is always assumed to be up). Tested and works. It is robust to the
wifi hotspot going up and down. It can reliably turn off and on the wifi at the
specified time intervals.